Update the API to use message operation store
This modifies chat extension to use the message operation store when
user calls add/edit/delete messages methods. Changes are immediately
cached to the message operation store. When operation is completed
successfully, the message operation is removed from store and
the updated message is updated in cache store (i.e. available for
fetches).

refs #156